+++ pkg/fleet/installer/msi/msilog.go
