+++ pkg/fleet/installer/msi/product.go
