+++ pkg/fleet/installer/msi/testdata/file_in_use.log
