+++ pkg/fleet/installer/msi/testdata/invalid_credentials.log
