+++ pkg/fleet/installer/msi/testdata/missing_password_for_dc.log
