+++ pkg/fleet/installer/msi/testdata/service_marked_for_deletion.log
