+++ pkg/fleet/installer/msi/testdata/wixfailwhendeferred.log
