+++ pkg/fleet/installer/packages/datadog_installer_windows.go
@@ -14,7 +14,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/DataDog/datadog-agent/pkg/fleet/internal/msi"
+	"github.com/DataDog/datadog-agent/pkg/fleet/installer/msi"
 	"github.com/DataDog/datadog-agent/pkg/fleet/internal/paths"
 )
